change to SmartLink object, move & rename fns, delete files, see #114
